Prepared package to deploy v0.1.3
Moved action_local_data_integrity.py out of the package since it's a local testing for now and added more documentation to it.

Cleaned execute_locally.py and fixed cross-references to action_local_data_integrity.py functions.

Added (actually it was in previous commits) configurations to __init__.py to make a custom deployment.